==> scripts/05_deliverable.Rmd <==
---
title: "Deliverable for Blue Water Baltimore"
author: "Eduardo Marin (with Dexter H Locke)"
date: "`r format(Sys.time())`"
output:
  html_document:
    theme: flatly
    code_folding: hide
    fig_width: 8
    fig_height: 7
    fig_caption: true
    toc: true
    toc_float: true
    self_contained: true
editor_options: 
  chunk_output_type: console
  markdown: 
    wrap: 72
---

# Blue Water Baltimore & Tree Giveaways

An entry-level analysis from the past ten+ years (2013 to 2024)

The document was prepared by Dexter H. Locke and Eduardo Marin. The research document will include the following sections with accompanying downloadable visuals for all the graphs.
- Location & Time (2/2)
- Provider (1/1)
- Species (1/1)
- Watershed (1/1)
- Demographics [JEDI & Tapestry] (0/1)
  - Income, Education, Race
  - Tapestry
- Origin-Destination (0/1)

```{r Loading Packages, message=FALSE, include=FALSE}
# Loading all packages

## List all of your packages here
knitr::opts_chunk$set(echo = TRUE)

packs <-c(
            'janitor'    # cleans things up, also pipe-friendly cross-tabulations
           , 'sf'         # for spatial data support
          , 'tidyverse'  # cuz
          # , 'tidylog'    # prints out what was done in dplyr and tidyr
          , 'magrittr'   # for the pipe
          , 'mapview'    # web maps for zooming and panning around
          #, 'beepr'      # makes noise when things are done!
          , 'tictoc'     # timing things.
          , 'raster'
          # , 'doParallel' # does what is says! PARALLEL
          # 'broom.mixed',# tidiers for mixed models AND nlme::gls()
          # , 'lubridate'   # DATES!
          , 'tidycensus' # tidy census package
          , 'tidygeocoder' # geo coding
          , 'leaflet' #creating the interactive mapping elements (more specific)
          , 'shiny'
          , 'leafsync'  # linked maps
          , 'RColorBrewer'
          , 'DT'
          , 'openxlsx'
          , 'mapdeck'
          , 'biscale' #used with ggplot to make bivariate maps
          , 'cowplot' #used to make more aesthetic ggplot visuals
          , 'ggspatial' #creates a basemap for your ggplots
          , 'classInt' #find me breaks without doing manualcalculations
          , 'rgl' #needs for rayshader
          , 'rayshader' #good data viz for 3d bar graphs
          , 'leaflet' #for heatmaps and more unique, customizable maps
          , 'leaflet.extras'
          , 'tidytext'
          )     

## IF the packages in 'packs' are not already installed, install them
## OTHERWISE do NOTHING
if (length(setdiff(packs, rownames(installed.packages()))) > 0) {
install.packages(setdiff(packs, rownames(installed.packages())))
}

## lapply(packs, library, character.only = TRUE)
## this actually loads them with library(package_name)
vapply(packs, library, character.only = TRUE, logical(1), logical.return = TRUE, quietly = TRUE)


## For tidycensus
## Setting for get_acs
census_api_key('58fc555c77c229747ade7d9fe50e7c71297cf91a', install = TRUE, overwrite = TRUE)
readRenviron("~/.Renviron")
options(tigris_use_cache = TRUE)
```

```{r Loading Census Data, include=FALSE}
# Pulling census data
## Listing needed variables
variables_acs <- 
  tidycensus::load_variables('acs5', year = 2021)

my_vars <- c(  'median_household_income' = 'B19013_001'
             , 'total_population' = 'B01003_001'
             , 'households' = 'B11001_001'
             , 'white_alone' = 'B03002_003'
             , 'median_age' = 'B01002_001'
             , 'total_population_over_25' = 'B15003_001'
             , 'ed_hs_degree' = 'B15003_017'
             , 'ed_ged_degree' = 'B15003_018'
             , 'ed_college_less_1_year' = 'B15003_019'
             , 'ed_college_more_1_year' = 'B15003_020'
             , 'ed_associate_degree' = 'B15003_021'
             , 'ed_plus_bachelor_degree' = 'B15003_022'
             , 'ed_plus_master_degree' = 'B15003_023'
             , 'ed_plus_professional_degree' = 'B15003_024'
             , 'ed_plus_doctorate_degree' = 'B15003_025'
            )
#TODO: Education is not divided by total population but by the education (confused with this and percentage)

(maryland <-
  tidycensus::get_acs(
      geography = 'state'
    , state = 'MD'
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
   st_transform(crs = st_crs(4326)))

(baltimore_city_counties <-
  tidycensus::get_acs(
      geography = 'county'
    , state = 'MD'
    , county = c('Baltimore City', 'Baltimore County')
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
   st_transform(crs = st_crs(4326))) 

(city <-
  tidycensus::get_acs(
      geography = 'county'
    , state = 'MD'
    , county = 'Baltimore City'
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
   st_transform(crs = st_crs(4326))) 

cbg <-
  tidycensus::get_acs(
      geography = 'block group'
    , state = 'MD'
    # , county = 'all'
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
    st_transform(crs = st_crs(4326)) |>
    mutate(jurisdiction = case_when(
      str_like(GEOID, '24510%') ~ "Baltimore City",
      str_like(GEOID, '24005%') ~ "Baltimore County",
      TRUE ~ "Neither"
    )) |>
    dplyr::select(-ends_with("M")) |> 
    rename_with(~ gsub("E$", "", .x), ends_with("E")) |>  
    rename(NAME = NAM) |>
    mutate(
      at_least_hs = rowSums(across(starts_with("ed_"))),
      at_least_college = rowSums(across(starts_with("ed_plus")))
    )|> #TODO: divided by total, multipled by 100 (from the base 25+ population) to get a percentage (rename these here)
    dplyr::select(-starts_with("ed_"))
```

```{r Loading Geographies, message=FALSE, include=FALSE}
# Reading in the shapefile and creating mapview layers
## Making the Maryland layer
state_layer <-
 maryland |>
 mapview(alpha.regions = 0, lwd = 2, col.regions = "#525252", alpha = 0.25, layer.name = "Maryland") 

## Making the Baltimore County layer
county_layer <-
  baltimore_city_counties |>
  mapview(alpha.regions = 0, lwd = 2, col.regions = "#525252", layer.name = "Baltimore City & County") 

## Making the Baltimore city layer
city_layer <-
  city |>
  rename("Name" = "NAME") |>
  mutate("Name" = "Baltimore city") |>
  mapview(col.regions = "#525252", alpha = 0.25, layer.name = "Baltimore City") 

city_layer <-

#Making the Baltimore neighborhood read
neighborhood_read <-
  st_read("../baltimore_tree_partners_data_to_big/shapefiles/Neighborhood/Neighborhood.shp") |>
  dplyr::select(`Name`,`geometry`,`Shape__Are`,`Shape__Len`) |>
  st_zm(drop = TRUE, what = "ZM") |>
  st_transform(4326) |>
  mutate(`Area (Square Miles)` = as.numeric(st_area(geometry)) / 2.59e+6)

## Making the Baltimore neighborhood layer
neighborhood_layer <-
  neighborhood_read |>
  mapview(alpha.regions = 0, lwd = 2, col.regions = "#525252", zcol = "Name", layer.name = "Baltimore's Neighborhood")

## Reading in the Blue Water Giveaway data
blue_water_giveaway_data <-
 st_read("../baltimore_tree_partners_data_to_big/shapefiles/bwb_giveaway_final/final_blue_water_baltimore_giveaways.shp")

## Making the Blue Water Giveaway layer
giveaway_layer <-
  blue_water_giveaway_data |>
  mapview(col.regions = "#5d6e75", alpha = 0.25, layer.name = "Total Tree Giveaways") 

## Fixing the Blue Water Giveaway Data  
neighborhood_intersect <-
  st_intersects(blue_water_giveaway_data, neighborhood_read, sparse = FALSE) |>
  apply(1, any) 

county_intersect <-
  st_intersects(blue_water_giveaway_data, baltimore_city_counties, sparse = FALSE) |>
  apply(1, any) 

blue_water_giveaway_data <-
  blue_water_giveaway_data |>
   mutate(county = case_when(
    neighborhood_intersect ~ "Baltimore City",
    county_intersect ~ "Baltimore County",
    TRUE ~ "Neither"
  ))

## Reading in the Watershed Data
blue_water_watershed_data <-
 st_read("../baltimore_tree_partners_data_to_big/shapefiles/Maryland_Watersheds_-_8_Digit_Watersheds/HYDR_Watersheds8Digit_DNR.shp") |>
 st_transform(4326) 

## Creating filter for the Watershed Data
watershed_to_keep <-  #DEXTER: From your changes to 04_exploring
  c('02130901' # Black river
    , '02130903' # Baltimore Harbor
    , '02130905' # Gwynns Falls
    , '02130904' # Jones Falls
    )

## Making the Watershed Layer
watershed_data <-
  blue_water_watershed_data |>
  filter(mde8digt %in% watershed_to_keep) |>
  dplyr::select(mde6digt,mde6name,mde8digt,mde8name,geometry) |>
  rename(`Watershed ID (6)` = mde6digt, `Watershed Name (6)` = mde6digt, `watershed` = mde8name, `watershed_id` = mde8digt) 

watershed_layer <-
  watershed_data |>
  mapview(col.regions = "#005273", alpha = 0.25, layer.name = "Watersheds")

#Adding this information to the BWB Data
blue_water_giveaway_data <-
  blue_water_giveaway_data |>
st_join(watershed_data, join = st_intersects) |>
  dplyr::select(-geometry)
```

```{r}
block_group_counts <-
  st_join(blue_water_giveaway_data, cbg[, c("NAME","GEOID")]) |>
  group_by(NAME.y, GEOID.y, year) |>
  summarise(count = n()) |>
  rename(NAME = NAME.y, GEOID = GEOID.y) |>
  st_drop_geometry()
  
census_count_map_layer <-
  cbg |>
  left_join(block_group_counts, by = c("NAME", "GEOID")) |>
  mutate(count = case_when(
    is.na(count) ~ 0,
    TRUE ~ count
  )) 

legend_order_census_count <- 
  classIntervals(
  census_count_map_layer$`count`,
  n = 6, # number of classes
  style = "fisher"
)$brks

census_count_map <-
  census_count_map_layer |>
  mapview(zcol = "count", layer.name = "Giveaway Count per Census Block", at = legend_order_census_count)
```

```{r Loading Tapestry, eval=FALSE, message=FALSE, include=FALSE}
## Making Tapestry
load('output_data/tap_lu_2024-06-25.RData')

tap_lu |> map(~levels(.)) 

load('output_data/bwb_total_giveaway_with_Tapestry_2019_2024-06-25.RData')

blue_water_giveaway_data_with_Tapestry_2019 |> 
  st_drop_geometry() |> 
  group_by(segment) |> 
  count() |> 
  ggplot(aes(n, segment)) +
  geom_col()
```

#Color Palette
```{r Color Palette, eval=FALSE, include=FALSE}
# List of Color Palette
##  (Earth Tones):
   "#f3f0e7"
   "#dbc5a5"
   "#cc9a86"
   "#bd806d"
   "#8c756a"

## Trees (Shades of Green):
   "#2f3456"
   "#5d6e75"
   "#8ba794"
   "#bac6b8"
   "#e9e4dc" 
   
## Branding (Shades of Blue)
   "#008CCC"
   "#005273"
   "#519342"
   
```
## Location & Time

In the past eleven years of running Blue Water Baltimore's giveaway
program, there has been a total of `r nrow(blue_water_giveaway_data)`
tree giveaways. Although the number is an overestimate since 2017 and
2018 included TreeBaltimore's values, the impact is undeniable. Blue Water's tree plantings have reached homeowners across the state of Maryland, from Frederick to Caroline County.

Figure 1. Giveaway Map of All Jurisdictions
```{r echo=FALSE}
giveaway_layer + state_layer #TODO: Fix the map zoom to focus on Maryland; Change the color of the trees to be different from the background (potentially make state darker)
```

Of course, most of these tree plantings have went to homeowners within the city of Baltimore and Baltimore County.

Figure 2. Giveaway Heatmap of All Jurisdiction
```{r}
leaflet(blue_water_giveaway_data) |>
  addProviderTiles("CartoDB.Positron") |>
  addHeatmap(lng = ~st_coordinates(blue_water_giveaway_data)[,1], 
             lat = ~st_coordinates(blue_water_giveaway_data)[,2], 
             blur = 5, 
             max = 0.10, 
             radius = 6) #add jurisdictions here (make a similar from watersheds)
                         #this one fixes with heatmap, just to show intensity
```

More specifically, these giveaways went to northern Baltimore County - near Towson - and northern Baltimore. If you solely look at Baltimore City and Baltimore County, you can see the following trend:

Figure 3. Map of Tree Giveaways per Census Block Group
```{r include=FALSE}
figure_3_census_count_layer <-
  cbg |>
  left_join(
  read_csv("output_data/census_count_year.csv")
    , by = "NAME"
  ) |>
  mutate(count = case_when(
    is.na(count) ~ 0,
    TRUE ~ count
  )) |>
  mutate(tree_per_household = count/households) |>
  filter(!is.na(tree_per_household)) #so the fischer doesn't break and try to divide by 0

figure_3_census_count_legend<- 
  classIntervals(
  figure_3_census_count_layer$`tree_per_household`,
  n = 5, 
  style = "fisher"
)$brks

figure_3_census_count_layer |>
mapview(zcol = "tree_per_household", layer.name = "Giveaway Count per Household", at = figure_3_census_count_legend)
```

Table 1. Table of Tree Plantings per Jurisdiction
```{r}
table_1_tree_plantings_per_jurisdiction <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(county) |>
  as_tibble() |>
  filter(county %in% c("Baltimore City", "Baltimore County", "Neither"))  |>
  rename(County = county, `Number of Giveaways` = n, Percent = percent) |>
  mutate(Percent = sprintf("%.2f%%", Percent * 100)) 

table_1_tree_plantings_per_jurisdiction |>
DT::datatable()
```

Graph 1. Graph of Tree Plantings per Jurisdiction per Year (Divided by Season)
```{r}
graph_1_season_colors <- c(
  "Fall" = "#008CCC",
  "Spring" = "#005273"
)

graph_1_tree_plantings_year_v1 <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  group_by(season, year) |>
  summarise(`Number of Giveaways` = n(), .groups = 'drop') |>
  rename(Season = season, Year = year) |>
  mutate(Year = as.double(Year)) |>
  ggplot(aes(x = Year, y = `Number of Giveaways`, fill = Season)) +
   geom_bar(position = "dodge", stat = "identity") +
    scale_fill_manual(values = graph_1_season_colors) +
    scale_y_continuous(
      limits = c(0, 2000),         
      breaks = seq(0, 2000, by = 250),  
      labels = scales::comma  
    ) +
    scale_x_continuous (
      breaks = seq(2013, 2023) 
    ) +
    labs(title = "Total Giveaways by Year and Season",
         x = "Year",
         y = "Total Giveaways",
         fill = "Season") +
    theme_minimal() 

graph_1_tree_plantings_year_v2 <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  group_by(season, year) |>
  summarise(`Number of Giveaways` = n(), .groups = 'drop') |>
  rename(Season = season, Year = year) |>
  mutate(Year = as.double(Year)) |>
  ggplot(aes(x = Year, y = `Number of Giveaways`, fill = Season)) +
   geom_bar(position = "stack", stat = "identity") +
    scale_fill_manual(values = graph_1_season_colors) +
    scale_y_continuous(
      limits = c(0, 2000),         
      breaks = seq(0, 2000, by = 250),  
      labels = scales::comma  
    ) +
    scale_x_continuous (
      breaks = seq(2013, 2023) 
    ) +
    labs(title = "Total Giveaways by Year and Season",
         x = "Year",
         y = "Total Giveaways",
         fill = "Season") +
    theme_minimal() 

# ggsave(
#   filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "location_graph_1_tree_plantings_general_v1.png"), #TODO: Fix to make more general file saving path
#   plot = graph_1_tree_plantings_year_v1,
#   width = 7,
#   height = 7
# ) #TODO: Season unstacked make for the graph, it shows what; Make vertical smaller since dodge (have it both ways)

graph_1_tree_plantings_year_v1

# ggsave(
#   filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "location_graph_1_tree_plantings_general_v2.png"), #TODO: Fix to make more general file saving path
#   plot = graph_1_tree_plantings_year_v2,
#   width = 7,
#   height = 7
# ) #TODO: Season unstacked make for the graph, it shows what; Make vertical smaller since dodge (have it both ways)

graph_1_tree_plantings_year_v2
```

The amount of tree giveaways that went to either Baltimore City or Baltimore County are nearly split halfway. `r table_1_tree_plantings_per_jurisdiction |> filter(County == "Baltimore City") |> pull(Percent)`of the tree giveaways went to Baltimore City and `r table_1_tree_plantings_per_jurisdiction |> filter(County == "Baltimore County") |> pull(Percent)` went to Baltimore County. However, the spread between city and county trees have shifted over the years. #TODO: Pull can be used completely instead of select

Table 2. Table of Tree Plantings per Jurisdiction per Year
```{r}
table_2_blue_water_baltimore_join <- 
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(county, year) |>
  as_tibble() |>
  filter(county %in% c("Baltimore City", "Baltimore County", "Neither"))

 table_2_blue_water_baltimore_join |>
 summarize(across(starts_with("20"), sum, na.rm = TRUE)) |>
 mutate(county = "Total") |>
 bind_rows(table_2_blue_water_baltimore_join) |>
 dplyr::select(11, 1:10) |>
 arrange(match(county, c(setdiff(unique(county), "Total"), "Total"))) |>
 as_tibble() |>
 rename(County = county) |>
 DT::datatable()
 
```

Graph 2. Graph of Tree Giveaways per Jurisdiction per Year (Divided by Season)
```{r}
graph_2_tree_plantings_year <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  group_by(season, year, county) |>
  summarise(`Number of Giveaways` = n(), .groups = 'drop') |>
  rename(Season = season, Year = year, County = county) |>
  mutate(Year = as.double(Year)) |>
  ggplot(aes(x = Year, y = `Number of Giveaways`, fill = Season)) +
   geom_bar(position = "stack", stat = "identity") +
    scale_fill_manual(values = graph_1_season_colors) +
    scale_y_continuous(
      limits = c(0, 2000),         
      breaks = seq(0, 2000, by = 250),  
      labels = scales::comma  
    ) +
    scale_x_continuous (
      breaks = seq(2013, 2023) 
    ) +
    labs(title = "Total Giveaways by Year and Season",
         x = "Year",
         y = "Total Giveaways",
         fill = "Season") +
    theme_minimal() +
    facet_wrap(~ County)

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "location_graph_2_tree_plantings_jurisdiction.png"), #TODO: Fix to make more general file saving path
  plot = graph_2_tree_plantings_year, 
  width = 21, 
  height = 7
)

graph_2_tree_plantings_year
```

The table below will highlight the amount per year and season but a larger scale, not bounded by any jurisctions, to understand the magnitude of tree giveaways over the year. 

Table 3. Table of Plantings in Each Neighborhood
```{r echo=FALSE}
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(year, season) |>
  as_tibble() |>
  add_row(year = "2020", Fall = 0, Spring = 0) |>
  mutate(Total = Fall + Spring) |>
  arrange(desc(Total)) |>
  rename(Year = year) |>
  DT::datatable()
```

The following section will take a more in depth look at the distribution of tree plantings in Baltimore, specifically at the neighborhood level. These will show the final destinations of those within the city:

Figure 4. Map of Tree Plantings in Baltimore Neighborhoods

```{r echo=FALSE}
figure_4_tree_neighborhoods_layer <-
neighborhood_read |>
  left_join(
    read_csv("output_data/neighborhood_count_bwb.csv")
    , by = "Name"
  ) |>
  mutate(count = case_when(
    is.na(count) ~ 0,
    TRUE ~ count
  )) |>
  rename(`Number of Trees` = count) |> #TODO(DEX): Unsure if we want to change this one as well to be based off of household instead.
  mutate(Density = `Number of Trees` / `Area (Square Miles)`) 

figure_4_tree_neighborhood_legend <- 
  classIntervals(
  figure_4_tree_neighborhoods_layer$`Density`,
  n = 10, 
  style = "fisher"
)$brks

figure_4_tree_neighborhoods_layer |>
  mapview(zcol = "Density", layer.name = "Giveaway Density (Tree per Square Mile)", at = figure_4_tree_neighborhood_legend) 
```

Table 4. Table of Plantings in Each Neighborhood
```{r echo=FALSE, message=FALSE}
 figure_4_tree_neighborhoods_layer |>
  st_drop_geometry() |>
  dplyr::select("Name","Density") |>
  arrange(desc(`Density`)) |>
  DT::datatable()
```

These may come surprising at first. However, these numbers are taking into consideration the amount of space within each jurisdiction. They are your number of trees in each neighborhood divided by the total square miles of each neighborhood. The following highlights the absolute value of tree plantings in each neighborhood. 

Table 5. Table of Plantings in Each Neighborhood
```{r echo=FALSE, message=FALSE}
figure_4_tree_neighborhoods_layer |>
  st_drop_geometry() |>
  dplyr::select("Name","Number of Trees") |>
  arrange(desc(`Number of Trees`)) |>
  DT::datatable()
```

Regardless, another interesting aspect to consider is how have the absolute number of tree plantings have changed over the years. The decision to not observe the density of tree plantings is since neighborhood boundaries can change, alongside the amount of plantable space over time.

Graph 3. Graph of Tree Plantings in Each Neighborhood Over the Years (Top 20)
```{r}
graph_3_filter_neighborhoods <-
  figure_4_tree_neighborhoods_layer |>
  dplyr::select(Name, `Number of Trees`) |>
  arrange(desc(`Number of Trees`)) |>
  slice(1:10)

graph_3_tree_plantings_year_neighborhood <-
figure_4_tree_neighborhoods_layer |>
  st_drop_geometry() |>
  left_join(
    read_csv("output_data/neighborhood_count_year_bwb.csv"), 
    by = "Name"
  ) |>
  rename(Year = year, Neighborhood = Name) |>
  dplyr::select(Neighborhood, `Number of Trees` = count, Year) |>
  dplyr::filter(Neighborhood %in% graph_3_filter_neighborhoods$Name) |>
  pivot_wider(names_from = Neighborhood, values_from = `Number of Trees`) |>
  add_row(Year = 2020) |>
  arrange((Year)) |>
  pivot_longer(cols = c("Ednor Gardens-Lakeside", "Hamilton", "Hampden", "Hamilton Hills", "Homeland", "Lauraville","Mount Washington","Roland Park","Westfield","Charles Village"), names_to = "Neighborhood", values_to = "Number of Trees") |>
  mutate(Neighborhood = factor(Neighborhood, levels = graph_3_filter_neighborhoods$Name)) |>
  ggplot(aes(x = Year, y = `Number of Trees`)) +
  geom_bar(position = "dodge", stat = "identity", fill = "#519342") +
  scale_x_continuous(breaks = seq(2013, 2023)) +
  coord_flip() +
  labs(title = "Top 10 Neighborhoods and Number of Trees", x = "Year", y = "Number of Trees") +
  theme_minimal() +
  facet_wrap(~ Neighborhood) 

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "location_graph_3_tree_plantings_neighborhoods_year.png"), #TODO: Fix to make more general file saving path
  plot = graph_3_tree_plantings_year_neighborhood, 
  width = 21, 
  height = 10
)

graph_3_tree_plantings_year_neighborhood

#TODO: Change the faceting to be read from left to right
```

As you can see, the top ten neighborhoods, in descending order, are `r paste(paste(graph_3_filter_neighborhoods$Name[1:9], collapse = ", "), ", and ", graph_3_filter_neighborhoods$Name[10], sep = "")`. The most surprising feature is that these neighborhoods vary greatly in terms of when their largest contribution were. Some, like Ednor Gardens-Lakeside, had the most giveaways in 2013, while others had them more recently like Hamilton Hills. 

Nevertheless, as important as it is to understand when and where these plantings have gone, it is also important to understand where they have been received.

## Provider

```{r include=FALSE}
unique_provider <-
 blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(gvwy_lc) |> 
  as_tibble() |> 
  arrange(desc(n)) |>
  slice(-1) 

unique_provider_na <- #TODO: Save out as a PNG
 blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(gvwy_lc) |> 
  as_tibble() |> 
  arrange(desc(n)) |>
  slice(1) 
```

In terms of giveaway location, there are approximately `r nrow(unique_provider)` different giveaway locations over the last ten years. Each of them vary, especially in terms of amount.

Table 6. Table of Giveaway Locations 
```{r echo=FALSE}
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(gvwy_lc) |> 
  rename(`Giveaway Location` = gvwy_lc) |>
  as_tibble() |> 
  arrange(desc(n)) |>
  slice(-1) |> 
  rename("Number of Trees" = n, Percent = percent) |>
  mutate(Percent = round(100 * Percent, 2)) |>
  dplyr::select(-`valid_percent`) |>
  DT::datatable()
```

In the case of the amount of tree giveaway locations, the most stems from Stillmeadow Community Fellowship. However, please understand that the largest majority are undefined, totaling `r unique(unique_provider_na$n)`.

Graph 4. Graph of Total Giveaway Locations 
```{r echo=FALSE}
graph_4_giveaway_locations <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(gvwy_lc) |>
  as_tibble() |> 
  arrange(desc(n)) |>
  slice(-1) |> 
  rename("Number of Trees" = n, `Giveaway Location` = gvwy_lc) |>
  mutate(percent = round(100 * percent, 2)) |>
  ggplot(aes(x = reorder(`Giveaway Location`, `Number of Trees`), y = `Number of Trees`)) +
    geom_col(fill = "#519342") +
    coord_flip() +
    labs(title = "Giveaway Location and Number of Trees", x = "Giveaway Location", y = "Number of Trees")
    theme_minimal() +
    NULL

# TODO: Fix the graph (currently offset)
# ggsave(
#   filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "provider_graph_4_tree_plantings_neighborhoods_year.png"), #TODO: Fix to make more general file saving path
#   plot = graph_4_giveaway_locations,
#   width = 21,
#   height = 10
# )
    
graph_4_giveaway_locations
```

The next section will begin prying into information concerning species from the tree giveaway. The goal will be to highlight new information while still continuing answering previous questions, like where have these species gone to spatially and what was the most sucessful species given per each provider. 

## Species

```{r include=FALSE}
colors_species <-  #TODO: Change colors to be more earth-tone like
  c(
  "Serviceberry" = "#9e0142",
  "Sweetbay Magnolia" = "#d53e4f", 
  "Pawpaw" = "#f46d43",
  "American Persimmon" = "#fdae61", 
  "Common Fig" = "#fee08b",
  "American Plum" = "#e6f598", 
  "American Holly" = "#abdda4",
  "Black Gum" = "#92b1e2",
  "Eastern Redbud" = "#3288bd", 
  "American Hornbeam" = "#5e4fa2"
)
```

The following map highlights the top ten species giveaway in the county.

Figure 5. Map of Top 10 Species Giveaway
```{r echo=FALSE}
blue_water_giveaway_data |>
  mutate(cmmn_nm = case_when(
    cmmn_nm %in% c("Serviceberry","Sweetbay Magnolia","Pawpaw","American Persimmon","Common Fig","American Plum","American Holly","Black Gum","Eastern Redbud", "American Hornbeam") ~    cmmn_nm, 
    TRUE ~ NA_character_
  )) |>
  filter(!is.na(cmmn_nm)) |>
  dplyr::select(year, season, cmmn_nm, gns_spc) |>
  mapview(zcol = "cmmn_nm", col.regions = colors_species, layer.name = "Top 10 Tree Species Planted from Blue Water Baltimore") #TODO: Fix how they are being highlighted and the order
```

```{r echo=FALSE}
#Prepping for the next data frame
datatable_6_giveaway <-
blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(cmmn_nm) |>
  rename(`Common Name` = cmmn_nm) |>
  arrange(desc(n)) |>
  mutate(percent = round(100 * percent, 2))
  
datatable_6_giveaway_top_10 <-
  datatable_6_giveaway |>
  slice(1:10)
```

As you can see, there is a fair distribution of the highest `r paste(datatable_6_giveaway_top_10$'Common Name'[1:9], collapse = ", ")`, and `r datatable_6_giveaway_top_10$'Common Name'[10]`. As this is shown in descending order, the most popular species for tree giveaways is `r datatable_6_giveaway_top_10$'Common Name'[1]`. Investing more time and energy into securing this species may be worthwhile for future tree giveaways.

The following table will show you more in-depth the breakdown of the different types of species given away.

Table 6. Table of the Species for the Giveaways
```{r echo=FALSE}
  datatable_6_giveaway |>
  rename(Count = n) |>
  DT::datatable()
```

In the case here, this visually displays the same information, highlighting how distinct the most common and least common species are in the amount of species give away.

Graph 5. Graph of the Most Common and Rare Tree Species
```{r echo=FALSE}
graph_5_common_rare_species <-
  datatable_6_giveaway |>
  rename(Count = n) |>
  mutate(group = rep(c("Common", "Rare"), each = 42)) |>
  ggplot(aes(Count, reorder(`Common Name`, Count))) + # use 'Count' instead of 'n'
  geom_col(fill = "#519342") +
  theme_bw(base_size = 10) +
  labs(y = 'Common Name',
       x = 'Number of trees',
       title = 'Which tree species were most and least common?') +
  facet_wrap(~group, scales = 'free') +
  NULL

# ggsave(
#   filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "species_graph_5_common_rare_species.svg"), #TODO: Fix to make more general file saving path
#   plot = graph_5_common_rare_species,
#   width = 22,
#   height = 10
# )

graph_5_common_rare_species
```

As mentioned earlier, an interesting aspect of this is to see and understand how these relate to prior topics. 

Table 7. Table of Species by Jurisdictions
```{r}
table_7_species_jurisdictions <-
blue_water_giveaway_data |>
  st_drop_geometry() |>
  rename(`Common Name` = cmmn_nm) |>
  group_by(`Common Name`,`county`) |>
  tabyl(`Common Name`,`county`) |>
  mutate(`Total` = rowSums(across(c(`Baltimore City`,`Baltimore County`, `Neither`)))) |>
  arrange(desc(Total)) 
  
table_7_species_jurisdictions |>
DT::datatable()

table_7_species_jurisdictions_top_5_city <-
  table_7_species_jurisdictions |>
  arrange(desc(`Baltimore City`)) |>
  slice(1:5)

table_7_species_jurisdictions_top_5_county <-
  table_7_species_jurisdictions |>
  arrange(desc(`Baltimore County`)) |>
  slice(1:5)
```

As you can see, for the most popular species, they tend to go more heavily towards Baltimore City. However, another unique aspect is finding out what the top five (5) of Baltimore City and Baltimore County are. Of course, you can just rearrange the data table. If you do not want to go through that process, however, here is the top five (5) for Baltimore City:

`r paste(paste(table_7_species_jurisdictions_top_5_city$'Common Name'[1:4], collapse = ", "), ", and ",table_7_species_jurisdictions_top_5_city$'Common Name'[5], sep = "")`.

In the case of the the top five (5) for Baltimore County, they are the following: 
`r paste(paste(table_7_species_jurisdictions_top_5_county$'Common Name'[1:4], collapse = ", "), ", and ",table_7_species_jurisdictions_top_5_county$'Common Name'[5], sep = "")`.

Of course, these change respective of time. The next part will underscore a graph of the top ten species within each jurisdiction: Baltimore City, Baltimore County, and neither. 

Graph 6. Graph of the Top Ten Species per each County
```{r}
graph_6_county_top_species <-
blue_water_giveaway_data |>
  st_drop_geometry() |>
  rename(`Common Name` = cmmn_nm) |>
  group_by(county, `Common Name`) |>
  summarize(Total = n(), .groups = 'drop') |>
  arrange(county, desc(Total)) |>
  group_by(county) |>
  slice_head(n = 10) |>
  group_by(county) |>
  mutate(`Common Name` = fct_reorder(`Common Name`, Total, .desc = TRUE)) |>
  ggplot(aes(x = `Common Name`, y = Total)) +
  geom_bar(stat = "identity", fill = "#519342") + #TODO: Fix the reordering to have it ascending/descending for both Baltimore County and Neither 
  coord_flip() +
  facet_wrap(~ county, scales = "free_y") +
  labs(title = "Top 10 Tree Species in Each County",
       x = "Common Name",
       y = "Total Number of Trees") +
  theme_minimal() 

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "species_graph_6_top_ten_species_jurisdiction.png"), #TODO: Fix to make more general file saving path
  plot = graph_6_county_top_species,
  width = 33,
  height = 10
)

graph_6_county_top_species 
```

Again, it is also important to understand what were the most popular species given
during each time of the year. 

Graph 7. Graph of the Top Ten Species per year
```{r}
graph_7_year_top_species <-
blue_water_giveaway_data |>
  st_drop_geometry() |>
  rename(`Common Name` = cmmn_nm) |>
  group_by(year, `Common Name`) |>
  summarize(Total = n(), .groups = 'drop') |>
  arrange(year, desc(Total)) |>
  group_by(year) |>
  slice_head(n = 10) |>
  mutate(`Common Name` = fct_reorder(`Common Name`, Total, .desc = TRUE)) |>
  ggplot(aes(x = `Common Name`, y = Total)) +
  geom_bar(stat = "identity", fill = "#519342") + #TODO: Fix the reordering to have it ascending/descending for both Baltimore County and Neither 
  coord_flip() +
  facet_wrap(~ year, scales = "free_y") +
  labs(title = "Top 10 Tree Species in Each Year",
       x = "Common Name",
       y = "Total Number of Trees") +
  theme_minimal() 

# ggsave(
#   filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "species_graph_7_top_ten_species_yearn.svg"), #TODO: Fix to make more general file saving path
#   plot = graph_7_year_top_species,
#   width = 20,
#   height = 15
# )

graph_7_year_top_species 
```

Finally, it is also important to understand if tree species differ based off of giveaway location. If so, what are the most popular tree species giveaway in the top five (5) giveaway location. The following graph, like the others, will attempt to highlight this, specifically the top 10 species in each one. 

Graph 8. Graph of the Top Ten Species per year
```{r}
top_five_giveaway_locations <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  rename(`Common Name` = cmmn_nm) |>
  group_by(gvwy_lc) |>
  summarize(Total = n(), .groups = 'drop') |>
  filter(!is.na(gvwy_lc)) |>
  top_n(5, Total) |>
  pull(gvwy_lc)

top_10_species_in_top_locations <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  rename(`Common Name` = cmmn_nm) |>
  filter(gvwy_lc %in% top_five_giveaway_locations) |>
  group_by(gvwy_lc, `Common Name`) |>
  summarize(Total = n(), .groups = 'drop') |>
  group_by(gvwy_lc) |>
  slice_head(n = 10) |>
  rename(`Giveaway Location` = gvwy_lc) |>
  group_by(`Common Name`,`Giveaway Location`) |>
  summarize(Total = sum(Total), .groups = 'drop') |>
  arrange(desc(Total)) |>
  mutate(`Common Name` = fct_reorder(`Common Name`, Total, .desc = TRUE)) |>
  ggplot(aes(x = `Common Name`, y = Total)) +
  geom_bar(stat = "identity", fill = "#519342") +
  coord_flip() +
  facet_wrap(~ `Giveaway Location`, scales = "free_y") +
  labs(title = "Top 10 Tree Species in the Top 5 Giveaway Locations",
       x = "Common Name",
       y = "Total Number of Trees") +
  theme_minimal() #TODO: Fix to make sure it is arranged in descending order

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "species_graph_top_10_species_in_top_locations.svg"), #TODO: Fix to make more general file saving path
  plot = top_10_species_in_top_locations,
  width = 40,
  height = 15
)

top_10_species_in_top_locations
```

TODO: Insert a blurb about the findings about but fix the descending order

In addition to species, information on where these tree giveaways have went relative to our watersheds is equally as important. 

## Watersheds

To begin, here is a general map of the watersheds in the Baltimore County and Baltimore City region.

Figure 6. Map of Watersheds in BWB Service Area (In Respective to the Counties)
```{r}
county_layer + watershed_layer
```

```{r}
#prepping for the statistic mentioned later
table_8_watershed_count <-
blue_water_giveaway_data |>
  st_drop_geometry() |>
  group_by(watershed) |>
  count() |>
  arrange(desc(n)) |>
  rename(Count = n) |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  ))
```

In terms of the plantings per watershed, these are the ones with the most ones in descending order: `r paste(paste(table_8_watershed_count$'Watershed Name'[1:15], collapse = ", "), ", and ",table_8_watershed_count$'Watershed Name'[16], sep = "")`. A special thing to notice is that there is a total of `r table_8_watershed_count |> filter('Watershed Name' == "Outside BWB Service Area") |> pull(Count)` outside of Blue Water Baltimore Service's area. The following will show how this, relatively speaking, is a far lower margin that what we might previously have imagined. 

Table 8. Table of Watersheds per Giveaway Count
```{r}
table_8_watershed_count_datatable <-
  blue_water_giveaway_data |>
  st_drop_geometry() |>
  tabyl(watershed) |>
  arrange(desc(n)) |>
  mutate(percent = round(100 * percent, 2)) |>
  dplyr::select(-valid_percent) |>
  rename(Count = n, Percent = percent) |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  rename(Watershed = watershed)

table_8_watershed_count_datatable |>
DT::datatable()
```

In terms of watersheds, it is important to see how many plants have gone to these areas over time. The following graph will highlight this and help distinguish if there is any statistical significance.

Graph 8. Number of Counts per Watershed per Year (per Season)
```{r}
graph_8_watersheds_count <-
  blue_water_giveaway_data |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  st_drop_geometry() |>
  rename(Watershed = watershed) |>
  group_by(year, Watershed, season) |>
  summarise(`Number of Giveaways` = n(), .groups = 'drop') |>
  rename(Year = year) |>
  mutate(Year = as.double(Year)) |>
  ggplot(aes(x = Year, y = `Number of Giveaways`, color = season, group = season)) +
  geom_line() + 
  geom_point() + 
  facet_wrap(~ `Watershed`, scales = "free_y") +
  scale_x_continuous(
    breaks = seq(2013, 2023, by = 2)
  ) +
  labs(title = "Total Giveaways by Year for Each Watershed",
       x = "Year",
       y = "Total Giveaways",
       color = "Season") +
  theme_minimal()
  
graph_8_watersheds_count 

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_seasons.png"), 
  plot = graph_8_watersheds_count,
  width = 40,
  height = 15
)

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_seasons.svg"), 
  plot = graph_8_watersheds_count,
  width = 40,
  height = 15
)
```

Here is a version of this but as a table.

Table 9. Table of Watersheds per Giveaway Count (per Season)
```{r}
table_9_watersheds_count_year <-
  blue_water_giveaway_data |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  st_drop_geometry() |>
  rename(Watershed = watershed) |>
  group_by(year, Watershed, season) |>
  summarise(`Number of Giveaways` = n(), .groups = "drop") |>
  pivot_wider(names_from = c("year","season"), values_from = `Number of Giveaways`) |>
  rename_with(~ str_replace_all(., "_", ", "))

table_9_watersheds_count_year |>
  DT::datatable()
```

Besides understanding the breakdown of each watershed by season, it is also important to understand the breakdown overall. In the following graph, it will highlight the total amount of plantings in each watershed per year.  

Graph 9. Number of Giveaway Counts per Watershed (per Year)
```{r}
graph_9_watersheds_count_year <-
  blue_water_giveaway_data |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  st_drop_geometry() |>
  rename(Watershed = watershed) |>
  group_by(season, year, Watershed) |>
  summarise(`Number of Giveaways` = n(), .groups = 'drop') |>
  rename(Year = year) |>
  mutate(Year = as.double(Year)) |>
  ggplot(aes(x = Year, y = `Number of Giveaways`)) +
  geom_col(fill = "#519342") + 
  scale_color_manual(values = graph_1_season_colors) +
  facet_wrap(~ `Watershed`, scales = "free_y") +
  scale_x_continuous(
    breaks = seq(2013, 2023, by = 2)
  ) +
  labs(title = "Total Giveaways by Year for Each Watershed",
       x = "Year",
       y = "Total Giveaways") +
  theme_minimal()
  
graph_9_watersheds_count_year 

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_year.png"), 
  plot = graph_9_watersheds_count_year,
  width = 40,
  height = 15
)

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_year.svg"), 
  plot = graph_9_watersheds_count_year,
  width = 40,
  height = 15
)
```

Table 10. Table of Watersheds per Giveaway Count (per Year)
```{r}
table_10_watersheds_count_year <-
  blue_water_giveaway_data |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  st_drop_geometry() |>
  rename(Watershed = watershed) |>
  group_by(year, Watershed) |>
  summarise(`Number of Giveaways` = n(), .groups = "drop") |>
  arrange(Watershed, `year`) |>
  pivot_wider(names_from = "year", values_from = `Number of Giveaways`)

table_10_watersheds_count_year |>
   DT::datatable()
```

Graph 10. Five (5) Popular Species per Watershed
```{r}
graph_10_watersheds_species_count <-
  blue_water_giveaway_data |>
  mutate(watershed = case_when(
    is.na(watershed) ~ "Outside BWB Service Area",
    TRUE ~ watershed
  )) |>
  st_drop_geometry() |>
  rename(Watershed = watershed, `Common Name` = cmmn_nm) |>
  group_by(Watershed, `Common Name`) |>
  summarise(`Number of Giveaways` = n(), .groups = "drop") |>
  arrange(Watershed, desc(`Number of Giveaways`)) |>
  group_by(Watershed) |>
  slice_head(n = 5) |>
  ggplot(aes(x = `Number of Giveaways`, y = `Common Name`)) +
  geom_col(fill = "#519342") + 
  facet_wrap(~ `Watershed`, scales = "free_y") +
  labs(title = "Total Giveaways by Species",
       x = "Total Giveaways",
       y = "Top Five (5) Species" ) +
  theme_minimal()
  
graph_10_watersheds_species_count 

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_top_five.png"), 
  plot = graph_10_watersheds_species_count,
  width = 40,
  height = 15
)

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_giveaways_watershed_top_five.svg"), 
  plot = graph_10_watersheds_species_count,
  width = 40,
  height = 15
)
```

As you can see here, the most common types of the 

## Distribution

We will first begin looking at income to see if there is any correlation between that and the number of giveaways. In terms of the scale, we will look specifically at the census block scale to determine this. We will first look at a synced map

Figure 6. Synced Map of Household Income and Giveaway Count (Per Census Blocks)
```{r}
## Making the legend breaks
legend_order_census_income_count <- 
  seq(0, 200000, length.out = 9) 

## Makiing the census-income map
census_income_map <-
  cbg |>
  rename("Median Household Income" = median_household_income) |>
  mutate(`Median Household Income` = case_when(
    is.na(`Median Household Income`) ~ 0,
    TRUE ~ `Median Household Income`)) |>
  mapview(zcol = "Median Household Income", layer.name = "Household Income and Census Block", at = legend_order_census_income_count)

## The Sync Map of Count and Income
sync(census_income_map, census_count_map)

#TODO: Double check if you should use a better color to visualize with the synced map
``` 

However, what may be more useful to see any patterns or trends is to look at a bivariate map. Bivariate maps work uniquely since they are on a color scale of two quantitative values. In the following case, [insert what the map states] 

Figure 7. Bivariate Map of Household Income and Giveaway Count (per Census Block)
```{r}
## Combining the
cbg_census_count <-
  cbg |>
  left_join(
  read_csv("output_data/census_count_year.csv")
    , by = "NAME"
  ) 

## setting the bivariate count and income
bivariate_count_income <-
  biscale::bi_class(cbg_census_count, x = median_household_income, y = count, style = "fisher", dim = 3)

## create the bivariate map
bivariate_count_income_map_pre <-
  bivariate_count_income |>
  ggplot() +
  geom_sf(mapping = aes(fill = bi_class), color = "white", 
          size = 0.1, show.legend = FALSE) +
  bi_scale_fill(pal = "PurpleOr", dim = 3) +
  labs(
    title = "Household Income and Giveaway Count",
    subtitle = "ACS & BWB Data"
  ) +
  theme(
    plot.title = element_text(size = 10),     
    plot.subtitle = element_text(size = 5)    
  ) +
  bi_theme()

## create the bivariate map's legend
bivariate_count_income_map_legend <- 
  bi_legend(pal = "PurpleOr",
            dim = 3,
            xlab = "Higher Median Household Income",
            ylab = "Higher Giveaway Count",
            size = 3)

## making combined map
bivariate_count_income_map <- 
  ggdraw() +
  draw_plot(bivariate_count_income_map_pre, 0, 0, 1, 1) +
  draw_plot(bivariate_count_income_map_legend, .25, .2, 0.2, 0.2)

## calling map
bivariate_count_income_map

ggsave(
  filename = file.path("/Users/eduardosmarin/Desktop/baltimore_tree_partners/deliverables", "graph_bivariate_income_giveaway.svg"), 
  plot = bivariate_count_income_map,
  width = 45,
  height = 12
)

```


```{r}
blue_water_giveaway_data |>
 mutate(household_income = cut(
    mdn_hs_,
    breaks = c(seq(0, 150000, by = 20000), Inf),
    labels = c(paste0("$", scales::comma(seq(0, 130000, by = 20000)), #scales::comma adds this to the number
                      " - $", scales::comma(seq(20000, 150000, by = 20000))), #scales::comma adds this to the number 
               "$150,000+"),
    include.lowest = TRUE,
    right = FALSE
  )) |>
  st_drop_geometry() |>
  group_by(household_income, year) |>
  summarise(`Number of Giveaways` = n()) |>
  view()


```










```{r}

legend_order_census_income_count <- 
  seq(0, 200000, length.out = 9) 

cbg_census_count |>
  ggplot(aes(x = year, y = median_household_income, fill = count)) +
  geom_tile() +
  scale_fill_gradientn(
    colours = c("blue", "red"), # Adjust colours as needed
    breaks = legend_order_census_income_count,
    labels = scales::comma(legend_order_census_income_count) # This adds commas to the labels for better readability
  ) +
  labs(x = "Year", y = "Income Bracket", fill = "Giveaway Counts") +
  theme_minimal()
```
