==> scripts/05_exploring_cleaned.Rmd <==
---
title: "Deliverable for Blue Water Baltimore"
author: "Eduardo Marin (with Dexter H Locke)"
date: "`r format(Sys.time())`"
output:
  html_document:
    theme: flatly
    code_folding: hide
    fig_width: 8
    fig_height: 7
    fig_caption: true
    toc: true
    toc_float: true
    self_contained: true
editor_options: 
  chunk_output_type: console
  markdown: 
    wrap: 72
---

```{r Loading Packages, message=FALSE, include=FALSE}
# Loading all packages

## List all of your packages here
knitr::opts_chunk$set(echo = TRUE)

packs <-c(
            'janitor'    # cleans things up, also pipe-friendly cross-tabulations
           , 'sf'         # for spatial data support
          , 'tidyverse'  # cuz
          # , 'tidylog'    # prints out what was done in dplyr and tidyr
          , 'magrittr'   # for the pipe
          , 'mapview'    # web maps for zooming and panning around
          #, 'beepr'      # makes noise when things are done!
          , 'tictoc'     # timing things.
          , 'raster'
          # , 'doParallel' # does what is says! PARALLEL
          # 'broom.mixed',# tidiers for mixed models AND nlme::gls()
          # , 'lubridate'   # DATES!
          , 'tidycensus' # tidy census package
          , 'tidygeocoder' # geo coding
          , 'leaflet' #creating the interactive mapping elements (more specific)
          , 'shiny'
          , 'leafsync'  # linked maps
          , 'RColorBrewer'
          , 'DT'
          , 'openxlsx'
          , 'mapdeck'
          , 'biscale' #used with ggplot to make bivariate maps
          , 'cowplot' #used to make more aesthetic ggplot visuals
          , 'ggspatial' #creates a basemap for your ggplots
          , 'classInt' #find me breaks without doing manualcalculations
          , 'rgl' #needs for rayshader
          , 'rayshader' #good data viz for 3d bar graphs
          , 'leaflet' #for heatmaps and more unique, customizable maps
          , 'leaflet.extras'
          , 'tidytext'
          )     

## IF the packages in 'packs' are not already installed, install them
## OTHERWISE do NOTHING
if (length(setdiff(packs, rownames(installed.packages()))) > 0) {
install.packages(setdiff(packs, rownames(installed.packages())))
}

## lapply(packs, library, character.only = TRUE)
## this actually loads them with library(package_name)
vapply(packs, library, character.only = TRUE, logical(1), logical.return = TRUE, quietly = TRUE)


## For tidycensus
## Setting for get_acs
census_api_key('58fc555c77c229747ade7d9fe50e7c71297cf91a', install = TRUE, overwrite = TRUE)
readRenviron("~/.Renviron")
options(tigris_use_cache = TRUE)
```


```{r Loading Tapestry, echo=TRUE, message=FALSE}
#ORIGINAL from 04_exploring,Rmd

#############################################################################################################
# This is needed to be able to correct the county data which is currently incorrect in the BWB Giveaway Data spreadsheet. The reason for this is probably since the spatial join was done incorrectly. 
##############################################################################################################

variables_acs <- 
  tidycensus::load_variables('acs5', year = 2021)

my_vars <- c(  'median_household_income' = 'B19013_001'
             , 'total_population' = 'B01003_001'
             , 'households' = 'B11001_001'
             , 'white_alone' = 'B03002_003'
             , 'median_age' = 'B01002_001'
             , 'total_population_over_25' = 'B15003_001'
             , 'ed_hs_degree' = 'B15003_017'
             , 'ed_ged_degree' = 'B15003_018'
             , 'ed_college_less_1_year' = 'B15003_019'
             , 'ed_college_more_1_year' = 'B15003_020'
             , 'ed_associate_degree' = 'B15003_021'
             , 'ed_plus_bachelor_degree' = 'B15003_022'
             , 'ed_plus_master_degree' = 'B15003_023'
             , 'ed_plus_professional_degree' = 'B15003_024'
             , 'ed_plus_doctorate_degree' = 'B15003_025'
            )

(baltimore_city_counties <- #Just need this to correct the shapefile
  tidycensus::get_acs(
      geography = 'county'
    , state = 'MD'
    , county = c('Baltimore City', 'Baltimore County')
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
   st_transform(crs = st_crs(4326))) |>
  rename(jursdiction = NAME) 

cbg <-
  tidycensus::get_acs(
      geography = 'block group'
    , state = 'Maryland'
    # , county = c('Baltimore City', 'Baltimore County')
    , variables = my_vars
    , year = 2021
    , geometry = TRUE
    , output = 'wide'
    , moe_level = 95
    ) |>
    st_transform(crs = st_crs(4326)) |>
    mutate(jurisdiction = case_when(
      str_like(GEOID, '24510%') ~ "Baltimore City",
      str_like(GEOID, '24005%') ~ "Baltimore County",
      TRUE ~ "Neither"
    )) |>
    dplyr::select(-ends_with("M")) |> 
    rename_with(~ gsub("E$", "", .x), ends_with("E")) |>  
    rename(NAME = NAM) |>
    mutate(
      pct_at_least_hs = (rowSums(across(starts_with("ed_"))))/total_population_over_25, #This gets percentage
      pct_at_least_college = (rowSums(across(starts_with("ed_plus"))))/total_population_over_25 #This gets percentage
    ) |>
    dplyr::select(-starts_with("ed_"))

#############################################################################################################
# This was supposed to be done in 02_double_checks. However, these, after the original meeting, were deemed incorrect or needed to be changed/added (which were done in the 04_explirng instead)
##############################################################################################################

species_data <- 
  tribble(
  ~common_name,        ~genus_species,            ~itree,
  "Hackberry",         "Celtis occidentalis",     "CEOC",
  "Willow Oak",        "Quercus phellos",         "QUPH",
  "Black Gum",         "Nyssa sylvatica",         "NYSY",
  "Eastern Redbud",    "Cercis canadensis",       "CECA",
  "Sassafras",         "Sassafras albidum",       "SAAL",
  "Chestnut Oak",      "Quercus prinus",          "QUPR",
  "Eastern Red Cedar", "Juniperus virginiana",    "JUVI",
  "Bald Cypress",      "Taxodium distichum",      "TADI",
  "Black Oak",         "Quercus velutina",        "QUVE",
  "American Holly",    "Ilex opaca",              "ILOP",
  "Sweetbay Magnolia", "Magnolia virginiana",     "MAVI",
  "American Sweetgum", "Liquidambar styraciflua", "LIST",
  "Bur Oak",           "Quercus macrocarpa",      "QUMA1",
  "Black Chokeberry",  "Aronia melanocarpa",      "ARME6",
  "Blackhaw Viburnum", "Viburnum prunifolium",    "VIPR",
  "Gingko",            "Ginkgo biloba",            "GIBI",
  "American Persimmon", "Diospyros virginiana",    "DIVI",
  "Flowering Cherry",   "Prunus yedoensis",        "PRYE",
  "Post Oak",           "Quercus stellata",        "QUST",
  "Chinkapin Oak",      "Quercus muehlenbergii",   "QUMU",
  "Tulip Poplar",       "Liriodendron tulipifera", "LITU",
  "Shortleaf Pine",     "Pinus echinata",          "PIEC",
  "Eastern White Pine", "Pinus americana",         "PIST",
  "Yellow Birch",       "Betula alleghaniensis",   "BEAL2",
  "American Yellowwood","Cladrastis luteau",       "CLLU",
  "Smooth Service Berry","Amelanchier laevis",     "AMLA",
  "Boxelder Maple",    "Acer negundo",             "ACNE",
  "American Linden",   "Tilia americana",          "TIAMRE",
  "American Plum",     "Prunus americana",         "PRAM"
  )
  
## loading data
blue_water_baltimore_data <-
  read_csv("output_data/bwb_giveaway_total_final_2024-05-13.csv") |>
  filter(!is.na(lat)) |> #Good double check
  st_as_sf(coords = c("long","lat"), crs = 4326) |> 
  st_join(cbg) |> 
  mutate(common_name = case_when(
    common_name == "Boxelder" ~ "Boxelder Maple",
    common_name == "Yoshino Flowering Cherry" ~ "Flowering Cherry",
    common_name == "Black Tupelo" ~ "Black Gum",
    common_name == "Chinquapin" ~ "Chinkapin Oak",
    common_name == "Breakfast" ~ "Other",
    common_name == "Blackhaw Vibernum" ~ "Blackhaw Viburnum",
    common_name == "Ginkgo" ~ "Gingko",
    TRUE ~ common_name
  )) |>
  filter(address != "822 Montpellier") |> 
  left_join(species_data, by = "common_name") |>
  mutate(
    genus_species = coalesce(genus_species.x, genus_species.y),
    itree = coalesce(itree.x, itree.y)
  ) |>
  dplyr::select(-genus_species.x, -genus_species.y, -itree.x, -itree.y) |>
  left_join(species_data, by = "genus_species") |>
  mutate(
    common_name = coalesce(common_name.x, common_name.y),
    itree = coalesce(itree.x, itree.y)
  ) |>
  dplyr::select(-common_name.x, -common_name.y, -itree.x, -itree.y, -county) |>
 mutate(genus_species = ifelse(common_name == "Other", "Other", genus_species),
       itree = ifelse(common_name == "Other", NA_character_, itree)) |>
  mutate(year = as.character(year)) |>
  mutate(giveaway_location = case_when(
    giveaway_location %in% c("Gwynns Falls/Leakin Park","Gwynns Falls Leakin Park", "GFLP") ~ "Gwynns Falls Leakin Park",
    TRUE ~ giveaway_location)) |>
  mutate(giveaway_location = case_when(
    giveaway_location == "Rogers Forge" ~ "Rodgers Forge",
    giveaway_location %in% c("West Towson","Westy Walk") ~ "West Towson ES",
    giveaway_location == "Gateway Park at Darley Park" ~ "Darley Park Gateway",
    giveaway_location == "Forestry Nursery" ~ "Herring Run Nursery",
    giveaway_location == "Greektown - John Ruhrah EMS" ~ "John Ruhrah ES",
    TRUE ~ giveaway_location
  ))

### Fixing the Blue Water Giveaway Data with the county as mentioned earlier

  blue_water_baltimore_data <-
  blue_water_baltimore_data |>
  st_join(baltimore_city_counties[, c("NAME")], join = st_intersects) |>
  rename(county = NAME.y, NAME = NAME.x) |>
  mutate(county = case_when(
    county == "Baltimore County, Maryland" ~ "Baltimore County",
    county == "Baltimore city, Maryland" ~ "Baltimore City",
    TRUE ~ "Neither"
  ))

#############################################################################################################
# This is new and is adding the watershed information. The issue right now which will show in your code will be the fact that your's will not work due to the file path. However, the way you have your listed in 04_exploring doesn't work for me since that file path creates an "error" on my end
##############################################################################################################

## Reading in the Watershed Data
blue_water_watershed_data <-
 st_read("../baltimore_tree_partners_data_to_big/shapefiles/Maryland_Watersheds_-_8_Digit_Watersheds/HYDR_Watersheds8Digit_DNR.shp") |> #DEXTER: You have something similar in your pull in the 04_exploring. However, when I try to do that, it does not work so I had to include the following "/baltimore_tree_partners_data_to_big/shapefiles/" to make it work. Potentially removing this on your end would make it work
 st_transform(4326)

## Creating filter for the Watershed Data
watershed_to_keep <-  #DEXTER: From your changes to 04_exploring
  c('02130901' # Black river
    , '02130903' # Baltimore Harbor
    , '02130905' # Gwynns Falls
    , '02130904' # Jones Falls
    , '02130906' # Patapsco River L N Br
    )

## Making the Watershed Layer
watershed_data <- 
  blue_water_watershed_data |>
  filter(mde8digt %in% watershed_to_keep) |>
  dplyr::select(mde6digt,mde6name,mde8digt,mde8name,geometry) |>
  rename(`Watershed ID (6)` = mde6digt, `Watershed Name (6)` = mde6digt, `watershed` = mde8name, `watershed_id` = mde8digt) 

#Adding this information to the BWB Data
blue_water_baltimore_data <-
  blue_water_baltimore_data |>
st_join(watershed_data, join = st_intersects) |>
  dplyr::select(-geometry) |>
  rename(geoid = GEOID, block_group_name = NAME)
```

```{r Loading Tapestry, echo=TRUE, message=FALSE}
#DEXTER: This is just removing the first `Watershed Name" and "ID" which isn't needed and the white_alone population. Besides that, it is making zip_code a character and reordering the columns to how you described in the email.
blue_water_baltimore_data_new <-
blue_water_baltimore_data |>
  dplyr::select(-`Watershed Name (6)`,-mde6name) |>
  mutate(zip_code = as.character(zip_code)) |>
  dplyr::select(year, season, date, genus_species, common_name, itree, address, city, state, jurisdiction, first_name, last_name, phone, email, giveaway_location, watershed_id, watershed, geoid, block_group_name, total_population, total_population_over_25, white_alone, median_household_income, median_age, pct_at_least_hs, pct_at_least_college, geometry) |>
  rename(itree_sp_code = itree) |>
  mutate(state = ifelse(is.na(state), "MD", state),
         watershed = ifelse(is.na(watershed), "Other", watershed),
         pct_at_least_hs = 100*(pct_at_least_hs),
         pct_at_least_college = 100*(pct_at_least_college)
         )

blue_water_baltimore_data_new |>
  st_write(paste0("output_data/blue_water_baltimore_data_", Sys.Date(), ".gpkg"), append = FALSE)

```


```{r}
# 1. Making other temporary table
temp_table <- 
  tribble(
  ~common_name,        ~genus_species,        ~itree_sp_code,
  "Black Walnut",      "Juglans nigra",         "FRPE",
  "Persimmon",         "Diospyros sp.",         "DIVI"
)


# 2. Makng the index table of species
index_table_species <- 
  blue_water_baltimore_data_new %>%
  st_drop_geometry() %>%
  dplyr::select(itree_sp_code, genus_species, common_name) %>%
  distinct() %>%
  dplyr::select(common_name, genus_species, itree_sp_code) %>%
  filter(!genus_species %in% c("prunus_americana", "Diospyros viginiana")) %>%
  mutate(common_name = if_else(common_name == "Serviceberry" &
                                 genus_species == "Amelanchier canadensis",
                               "Serviceberry (CAN)",
                               common_name)) %>%
  bind_rows(temp_table)
 

# 3. Reading the BWB data
pre_bwb_data_2024 <- 
  st_read("/Users/eduardosmarin/Desktop/baltimore_tree_partners/input_data/BWB_2024_TreeGiveaways.gpkg") |>
   mutate(common_name = str_to_title(common_name)) |>
  mutate(common_name = case_when(
    common_name == "Hop Hornbeam" ~ "American Hophornbeam",
    common_name == "Witch Hazel" ~ "American Witchhazel",
    common_name == "Canadian Serviceberry" ~ "Serviceberry (CAN)",
    common_name == "Allegheny Serviceberry" ~ "Smooth Service Berry",
    TRUE ~ common_name
  )) |>
  dplyr::select(-genus_species, -itree_sp_code) |>
  left_join(index_table_species, by = "common_name") |>
dplyr::select(-watershed_id         
, -watershed               
, -geoid                   
, -block_group_name        
, -total_population        
, -total_population_over_25
, -white_alone             
, -median_household_income 
, -median_age              
, -pct_at_least_hs         
, -pct_at_least_college )


# 4. Adding the watershed and Census Data
cbg_renamed <- 
  cbg |>
  rename(cbg_geoid = GEOID, cbg_name = NAME)

watershed_renamed <- 
  watershed_data |>
  rename(watershed_name = `Watershed Name (6)`)

bwb_data_2024 <- 
  pre_bwb_data_2024 |>
  st_join(cbg_renamed, join = st_intersects) |>
  st_join(watershed_renamed, join = st_intersects) |>
  dplyr::select(-jurisdiction.y) |>
  rename(jurisdiction = jurisdiction.x) |>
  mutate(wateshed = ifelse(is.na(watershed), "Other", watershed)) |>
  rename(geoid = cbg_geoid
       , block_group_name = cbg_name) |>
  rename(geom = Shape) |>
  dplyr::select(year                
, season                  
, date                    
, genus_species           
, common_name             
, itree_sp_code           
, address                 
, city                    
, state                   
, jurisdiction
, first_name              
, last_name               
, phone                   
, email                   
, giveaway_location       
, watershed_id            
, watershed               
, geoid                   
, block_group_name        
, total_population        
, total_population_over_25
, white_alone             
, median_household_income 
, median_age              
, pct_at_least_hs         
, pct_at_least_college
, geom
  ) |>
  mutate(
    phone_clean = str_replace_all(phone, "[^0-9]", ""),
    phone_number = case_when(
      nchar(phone_clean) == 10 ~ str_replace(phone_clean, "(\\d{3})(\\d{3})(\\d{4})", "\\1-\\2-\\3"),
      nchar(phone_clean) == 11 & str_starts(phone_clean, "1") ~ str_replace(substr(phone_clean, 2, 11), "(\\d{3})(\\d{3})(\\d{4})", "\\1-\\2-\\3"),
      TRUE ~ NA_character_
    )
  ) |>
  dplyr::select(-phone, -phone_clean) |>
  rename(phone = phone_number)


# 6. Reading the entire year data 
bwb_all <-
  st_read("../baltimore_tree_partners/output_data/blue_water_baltimore_data_2024-12-27.gpkg")

# 6. Combinig and writing out the new one
bwb_giveaway_data_2013_2024 <-
  rbind(bwb_all, bwb_data_2024) |>
  mutate(genus_species = case_when(
    genus_species == "prunus_americana" ~ "Prunus americana",
  TRUE ~ genus_species
  )) |> 
  mutate(jurisdiction = str_to_title(jurisdiction)) |>
  mutate(jurisdiction = ifelse(jurisdiction == "Carroll County", "Neither", jurisdiction)) |>
  mutate(genus_species = ifelse(genus_species == "Diospyros viginiana", "Diospyros virginiana", genus_species)) |>
mutate(genus_species = ifelse(genus_species == "Gymnocladus dioicus", "Gymnocladus dioicusa", genus_species)) |>
  mutate(genus_species = ifelse(genus_species == "Cladrastis kentuckea", "	
Cladrastis kentuckea", genus_species)) |>
  mutate(genus_species = ifelse(common_name == "Serviceberry (CAN)", "Serviceberry", genus_species)) |>
  mutate(city = str_to_title(city)) |>
  mutate(city = case_when(
    city == "Elliott City" ~ "Ellicott City"
  , city == "Eldridge" ~ "Elkridge"
  , city == "Lutherville Timonium" ~ "Lutherville-Timonium"
  , city == "Randalstown" ~ "Randallstown"
  , city %in% c("Reisterstown", "Reisterstwon", "Reistertown") ~ "Reisterstown"
  , TRUE ~ city
  )) |> 
 dplyr::select(-jurisdiction) |>
 st_join(baltimore_city_counties) |>
 dplyr::select(-GEOID
, -median_household_incomeE     
, -median_household_incomeM     
, -total_populationE            
, -total_populationM            
, -householdsE                  
, -householdsM                  
, -white_aloneE                 
, -white_aloneM                 
, -median_ageE                  
, -median_ageM                  
, -total_population_over_25E    
, -total_population_over_25M    
, -ed_hs_degreeE                
, -ed_hs_degreeM                
, -ed_ged_degreeE               
, -ed_ged_degreeM               
, -ed_college_less_1_yearE      
, -ed_college_less_1_yearM      
, -ed_college_more_1_yearE      
, -ed_college_more_1_yearM      
, -ed_associate_degreeE         
, -ed_associate_degreeM         
, -ed_plus_bachelor_degreeE     
, -ed_plus_bachelor_degreeM     
, -ed_plus_master_degreeE       
, -ed_plus_master_degreeM       
, -ed_plus_professional_degreeE 
, -ed_plus_professional_degreeM 
, -ed_plus_doctorate_degreeE    
, -ed_plus_doctorate_degreeM) |>
  mutate(jurisdiction = case_when(
    NAME == "Baltimore County, Maryland" ~ "Baltimore County"
  , NAME == "Baltimore city, Maryland" ~ "Baltimore City"
  , TRUE ~ "Other"
  )) |>
  dplyr::select(-NAME)

# 7. FINAL SAVE OUT :D
bwb_giveaway_data_2013_2024 |>
  filter(year == "2024") |>
  st_write(paste0("output_data/year_2024_bwb_giveaway_", Sys.Date(), ".gpkg"), append = FALSE)

bwb_giveaway_data_2013_2024 |>
  st_write(paste0("output_data/year_2013_to_2024_bwb_giveaway_", Sys.Date(), ".gpkg"), append = FALSE)
  
```


```{r}
# 1. Within residential, front and back comparison
# 2. Cross city comparison,
# 3. Is the gap in survivorship biger or lower in higher-income and 
# 4. Summarize the neighborhood level
# 5. Do higher income neighborhoods smaller gap wolevel gap
# 6. Do higher income neighborhoods smaller 
# 7. A third level question that will be constriained
# 8. Double check the geopackage.... :D :;-;-;-;-; |P

```




```{r Loading Tapestry, eval=FALSE, message=FALSE, include=FALSE}
blue_water_baltimore_data_check <- 
  bwb_giveaway_data_2013_2024
  # st_read("output_data/blue_water_baltimore_data_2024-08-23.gpkg") #DEXTER: Change whenver your sys.date does it

blue_water_baltimore_data_check |> glimpse()
blue_water_baltimore_data_check |> summary()
blue_water_baltimore_data_check |> map(~levels(.))
blue_water_baltimore_data_check |> mapview(zcol = 'median_household_income')

# more scruitiny
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(year)
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(year, season)
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(date)
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(genus_species)

species_checker <- 
  blue_water_baltimore_data_check |> 
  st_drop_geometry() |> 
  tabyl(genus_species, common_name) |> 
  tibble() |> 
  pivot_longer(-genus_species) |> 
  filter(value > 0)

itree_species_checker <- 
  blue_water_baltimore_data_check |> 
  st_drop_geometry() |> 
  tabyl(genus_species, itree_sp_code) |> 
  tibble() |> 
  pivot_longer(-genus_species) |> 
  filter(value > 0)

# should be NONE! wohoo
species_checker |> anti_join(itree_species_checker, by = 'genus_species')

# should be NONE! wohoo
species_checker |> 
  left_join(itree_species_checker, by = 'genus_species') |> 
  filter(value.x != value.y)


blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(city, state)

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(city, jurisdiction)

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(phone) # is this right?
blue_water_baltimore_data_check |> st_drop_geometry() |> group_by(phone, first_name) |> count()|> arrange(phone)# is this right? are these multiple trees per person/phone number?

#DEXTER: assuming this could be because many people have the same first name, trying to see if you group by both first and last if it differs
blue_water_baltimore_data_check |> st_drop_geometry() |> group_by(phone, first_name, last_name, year) |> count()|> arrange(desc(n)) |> view()

#DEXTER: Did the first five checks of the original spreadsheets and seem to line up. It is weird though that there are some with like ~20 to ~30 trees for one person in one year. This could be a question for Darin
# TODO Eduardo, check the original spreadsheets? 
#DEXTER: I checked those for the first five there present and they lined up. It is very weird though and would suggest asking darin

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(giveaway_location) # TODO missing locations ok? 
#DEXTER: 

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(giveaway_location, year)
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(year, season, giveaway_location) |> bind_rows(.id = 'loc')

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(watershed_id, watershed) # change NA to "other" in "watershed" column

blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(geoid)
blue_water_baltimore_data_check |> st_drop_geometry() |> tabyl(geoid, jurisdiction)

# very good
blue_water_baltimore_data_check %$% cor.test(total_population, total_population_over_25)

# TODO multiple education by 100 to make percent, and add "pct_" to column name (COMPLETED)
blue_water_baltimore_data_check %$% cor.test(pct_at_least_hs, pct_at_least_college)

blue_water_baltimore_data_check %$% cor.test(median_household_income, pct_at_least_college) # good.

blue_water_baltimore_data_check |> 
  ggplot(aes(median_household_income)) + 
  geom_density() # good

############  
```
